requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
mysql-connector-python==8.2.0
pytest-xdist==3.3.1
